requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
numpy
pillow
python-multipart